@@ -3,40 +3,42 @@
   "version": "0.0.1",
   "private": true,
   "scripts": {
-    "start": "node node_modules/react-native/local-cli/cli.js start",
+    "start": "react-native start",
     "test": "jest",
     "lint": "eslint ."
   },
   "dependencies": {
-    "@react-native-community/async-storage": "^1.5.0",
-    "mobx": "^5.10.1",
+    "@react-native-community/async-storage": "^1.6.1",
+    "mobx": "^5.13.0",
     "mobx-react": "^5.4.4",
-    "mobx-state-tree": "^3.14.0",
-    "react": "16.8.6",
-    "react-native": "0.59.9",
+    "react": "^16.9.0",
+    "react-native": "0.60.4",
     "react-native-gesture-handler": "^1.3.0",
     "react-native-hyperlink": "^0.0.14",
-    "react-native-image-picker": "^0.28.1",
-    "react-native-linear-gradient": "^2.5.4",
+    "react-native-image-picker": "^1.0.2",
+    "react-native-linear-gradient": "^2.5.6",
     "react-native-picker": "^4.3.7",
-    "react-native-vector-icons": "^6.5.0",
-    "react-navigation": "^3.11.0"
+    "react-native-vector-icons": "^6.6.0",
+    "react-navigation": "^3.11.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.4.5",
-    "@babel/plugin-proposal-class-properties": "^7.4.4",
+    "@babel/core": "^7.5.5",
+    "@babel/plugin-proposal-class-properties": "^7.5.5",
     "@babel/plugin-proposal-decorators": "^7.4.4",
     "@babel/plugin-transform-flow-strip-types": "^7.4.4",
-    "@babel/runtime": "^7.4.5",
+    "@babel/runtime": "^7.5.5",
     "@react-native-community/eslint-config": "^0.0.5",
-    "@types/react": "^16.8.22",
-    "@types/react-native": "^0.57.63",
-    "@types/react-navigation": "^3.0.7",
+    "@types/jest": "^24.0.17",
+    "@types/react": "^16.9.1",
+    "@types/react-native": "^0.60.3",
+    "@types/react-navigation": "^3.0.8",
+    "@types/react-test-renderer": "^16.9.0",
     "babel-jest": "^24.8.0",
-    "eslint": "^6.0.1",
+    "eslint": "^6.1.0",
     "jest": "^24.8.0",
-    "metro-react-native-babel-preset": "^0.54.1",
-    "react-test-renderer": "16.8.6",
+    "metro-react-native-babel-preset": "^0.55.0",
+    "react-native-typescript-transformer": "^1.2.13",
+    "react-test-renderer": "^16.9.0",
     "tslint": "^5.18.0",
     "typescript": "^3.5.2"
   },
